+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 #### FIXES
  - `icwt` with `scaletype='linear'`: fix constant scaling factor
+ - `trapz` -> `trapezoid`; scipy deprecation, `scipy.integrate.trapz`
+ - numpy deprecation, fix `int()` upon 1D array (of size 1)
 
 ### 0.6.5